DC-1874: Modified Node to include FileInfo directory.
